Various small code cleanups: - Lots of "signed vs unsigned" comparison fixes - time_t globally replaced with TIME_T - All seqnos moved to "seqno_t", which defaults to unsigned long - DIMOF() definition centralized to portability.h and typecast to int - EOS define moved to portability.h - dropped inclusion of signal.h from stonith.h, so that sigignore is   properly defined
CVS patchset: 1584
CVS date: 2003/01/31 10:02:09

+++ lib/clplumbing/ipcsocket.c
@@ -583,7 +583,7 @@ socket_resume_io(struct IPC_CHANNEL *ch)
       }
     
       if (len > 0 ) {
-	if (len < msg->msg_len){
+	if ((size_t)len < msg->msg_len){
 	  /* 
 	   * FIXME! for stream domain socket, if the message is too big, it 
 	   * may cause part of the message cut instead of being sent out.
